# Code changes to prepare for JUCE upgrade, speed up builds