--- a/src/AudioBufferSource.h
+++ b/src/AudioBufferSource.h
@@ -13,30 +13,28 @@
 #ifndef OPENSHOT_AUDIOBUFFERSOURCE_H
 #define OPENSHOT_AUDIOBUFFERSOURCE_H
 
-#include <iomanip>
-#include <OpenShotAudio.h>
+#include <AppConfig.h>
+#include <juce_audio_basics/juce_audio_basics.h>
 
-/// This namespace is the default namespace for all code in the openshot library
 namespace openshot
 {
-
 	/**
-	 * @brief This class is used to expose an AudioSampleBuffer as an AudioSource in JUCE.
+	 * @brief This class is used to expose an AudioBuffer<float> as an AudioSource in JUCE.
 	 *
-	 * The <a href="http://www.juce.com/">JUCE</a> library cannot play audio directly from an AudioSampleBuffer, so this class exposes
-	 * an AudioSampleBuffer as a AudioSource, so that JUCE can play the audio.
+	 * The <a href="http://www.juce.com/">JUCE</a> library cannot play audio directly from an AudioBuffer<float>, so this class exposes
+	 * an AudioBuffer<float> as a AudioSource, so that JUCE can play the audio.
 	 */
 	class AudioBufferSource : public juce::PositionableAudioSource
 	{
 	private:
 		int position;
 		bool repeat;
-		juce::AudioSampleBuffer *buffer;
+		juce::AudioBuffer<float> *buffer;
 
 	public:
 		/// @brief Default constructor
 		/// @param audio_buffer This buffer contains the samples you want to play through JUCE.
-		AudioBufferSource(juce::AudioSampleBuffer *audio_buffer);
+		AudioBufferSource(juce::AudioBuffer<float> *audio_buffer);
 
 		/// Destructor
 		~AudioBufferSource();
@@ -69,7 +67,7 @@ namespace openshot
 		void setLooping (bool shouldLoop);
 
 		/// Update the internal buffer used by this source
-		void setBuffer (juce::AudioSampleBuffer *audio_buffer);
+		void setBuffer (juce::AudioBuffer<float> *audio_buffer);
 	};
 
 }