--- a/src/AudioResampler.h
+++ b/src/AudioResampler.h
@@ -14,7 +14,10 @@
 #define OPENSHOT_RESAMPLER_H
 
 #include "AudioBufferSource.h"
-#include <OpenShotAudio.h>
+
+#include <AppConfig.h>
+#include <juce_audio_basics/juce_audio_basics.h>
+#include <juce_audio_devices/juce_audio_devices.h>
 
 namespace openshot {
 
@@ -26,8 +29,8 @@ namespace openshot {
 	 */
 	class AudioResampler {
 	private:
-		juce::AudioSampleBuffer *buffer;
-		juce::AudioSampleBuffer *resampled_buffer;
+		juce::AudioBuffer<float> *buffer;
+		juce::AudioBuffer<float> *resampled_buffer;
 		openshot::AudioBufferSource *buffer_source;
 		juce::ResamplingAudioSource *resample_source;
 		juce::AudioSourceChannelInfo resample_callback_buffer;
@@ -49,15 +52,15 @@ namespace openshot {
 		/// @param new_buffer The buffer of audio samples needing to be resampled
 		/// @param sample_rate The original sample rate of the buffered samples
 		/// @param new_sample_rate The requested sample rate you need
-		void SetBuffer(juce::AudioSampleBuffer *new_buffer, double sample_rate, double new_sample_rate);
+		void SetBuffer(juce::AudioBuffer<float> *new_buffer, double sample_rate, double new_sample_rate);
 
 		/// @brief Sets the audio buffer and key settings
 		/// @param new_buffer The buffer of audio samples needing to be resampled
 		/// @param ratio The multiplier that needs to be applied to the sample rate (this is how resampling happens)
-		void SetBuffer(juce::AudioSampleBuffer *new_buffer, double ratio);
+		void SetBuffer(juce::AudioBuffer<float> *new_buffer, double ratio);
 
 		/// Get the resampled audio buffer
-		juce::AudioSampleBuffer* GetResampledBuffer();
+		juce::AudioBuffer<float>* GetResampledBuffer();
 	};
 
 }