--- a/src/CVStabilization.cpp
+++ b/src/CVStabilization.cpp
@@ -16,6 +16,7 @@
 #include <iostream>
 
 #include "CVStabilization.h"
+#include "Exceptions.h"
 
 #include "stabilizedata.pb.h"
 #include <google/protobuf/util/time_util.h>