--- a/src/CVTracker.cpp
+++ b/src/CVTracker.cpp
@@ -20,6 +20,7 @@
 #include "OpenCVUtilities.h"
 #include "CVTracker.h"
 #include "trackerdata.pb.h"
+#include "Exceptions.h"
 
 using namespace openshot;
 using google::protobuf::util::TimeUtil;