--- a/src/CacheDisk.cpp
+++ b/src/CacheDisk.cpp
@@ -81,7 +81,7 @@ void CacheDisk::CalculateRanges() {
 	if (needs_range_processing) {
 
 		// Create a scoped lock, to protect the cache from multiple threads
-		const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+                const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 		// Sort ordered frame #s, and calculate JSON ranges
 		std::sort(ordered_frame_numbers.begin(), ordered_frame_numbers.end());
@@ -139,16 +139,15 @@ CacheDisk::~CacheDisk()
 	frame_numbers.clear();
 	ordered_frame_numbers.clear();
 
-	// remove critical section
-	delete cacheCriticalSection;
-	cacheCriticalSection = NULL;
+	// remove mutex
+	delete cacheMutex;
 }
 
 // Add a Frame to the cache
 void CacheDisk::Add(std::shared_ptr<Frame> frame)
 {
 	// Create a scoped lock, to protect the cache from multiple threads
-	const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+	const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 	int64_t frame_number = frame->number;
 
 	// Freshen frame if it already exists
@@ -207,7 +206,7 @@ void CacheDisk::Add(std::shared_ptr<Frame> frame)
 std::shared_ptr<Frame> CacheDisk::GetFrame(int64_t frame_number)
 {
 	// Create a scoped lock, to protect the cache from multiple threads
-	const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+	const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 	// Does frame exists in cache?
 	if (frames.count(frame_number)) {
@@ -277,7 +276,7 @@ std::shared_ptr<Frame> CacheDisk::GetFrame(int64_t frame_number)
 std::shared_ptr<Frame> CacheDisk::GetSmallestFrame()
 {
 	// Create a scoped lock, to protect the cache from multiple threads
-	const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+	const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 	// Loop through frame numbers
 	std::deque<int64_t>::iterator itr;
@@ -300,7 +299,7 @@ std::shared_ptr<Frame> CacheDisk::GetSmallestFrame()
 int64_t CacheDisk::GetBytes()
 {
 	// Create a scoped lock, to protect the cache from multiple threads
-	const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+	const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 	int64_t  total_bytes = 0;
 
@@ -322,7 +321,7 @@ void CacheDisk::Remove(int64_t frame_number)
 void CacheDisk::Remove(int64_t start_frame_number, int64_t end_frame_number)
 {
 	// Create a scoped lock, to protect the cache from multiple threads
-	const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+	const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 	// Loop through frame numbers
 	std::deque<int64_t>::iterator itr;
@@ -374,7 +373,7 @@ void CacheDisk::MoveToFront(int64_t frame_number)
 	if (frames.count(frame_number))
 	{
 		// Create a scoped lock, to protect the cache from multiple threads
-		const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+		const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 		// Loop through frame numbers
 		std::deque<int64_t>::iterator itr;
@@ -397,7 +396,7 @@ void CacheDisk::MoveToFront(int64_t frame_number)
 void CacheDisk::Clear()
 {
 	// Create a scoped lock, to protect the cache from multiple threads
-	const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+	const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 	// Clear all containers
 	frames.clear();
@@ -418,7 +417,7 @@ void CacheDisk::Clear()
 int64_t CacheDisk::Count()
 {
 	// Create a scoped lock, to protect the cache from multiple threads
-	const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+	const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 	// Return the number of frames in the cache
 	return frames.size();
@@ -431,7 +430,7 @@ void CacheDisk::CleanUp()
 	if (max_bytes > 0)
 	{
 		// Create a scoped lock, to protect the cache from multiple threads
-		const GenericScopedLock<CriticalSection> lock(*cacheCriticalSection);
+		const std::lock_guard<std::recursive_mutex> lock(*cacheMutex);
 
 		while (GetBytes() > max_bytes && frame_numbers.size() > 20)
 		{