--- a/src/ChannelLayouts.h
+++ b/src/ChannelLayouts.h
@@ -57,7 +57,6 @@ enum ChannelLayout
 };
 
 
-
 }
 
 #endif