--- a/src/Clip.cpp
+++ b/src/Clip.cpp
@@ -11,6 +11,8 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
 #include "Clip.h"
+
+#include "AudioResampler.h"
 #include "Exceptions.h"
 #include "FFmpegReader.h"
 #include "FrameMapper.h"
@@ -453,13 +455,13 @@ std::string Clip::get_file_extension(std::string path)
 }
 
 // Reverse an audio buffer
-void Clip::reverse_buffer(juce::AudioSampleBuffer* buffer)
+void Clip::reverse_buffer(juce::AudioBuffer<float>* buffer)
 {
 	int number_of_samples = buffer->getNumSamples();
 	int channels = buffer->getNumChannels();
 
 	// Reverse array (create new buffer to hold the reversed version)
-	juce::AudioSampleBuffer *reversed = new juce::AudioSampleBuffer(channels, number_of_samples);
+	juce::AudioBuffer<float> *reversed = new juce::AudioSampleBuffer(channels, number_of_samples);
 	reversed->clear();
 
 	for (int channel = 0; channel < channels; channel++)
@@ -491,10 +493,10 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 	// Check for a valid time map curve
 	if (time.GetLength() > 1)
 	{
-		const juce::GenericScopedLock<juce::CriticalSection> lock(getFrameCriticalSection);
+		const std::lock_guard<std::recursive_mutex> lock(getFrameMutex);
 
 		// create buffer and resampler
-		juce::AudioSampleBuffer *samples = NULL;
+		juce::AudioBuffer<float> *samples = NULL;
 		if (!resampler)
 			resampler = new AudioResampler();
 
@@ -514,10 +516,10 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 			if (time.GetRepeatFraction(frame_number).den > 1) {
 				// SLOWING DOWN AUDIO
 				// Resample data, and return new buffer pointer
-				juce::AudioSampleBuffer *resampled_buffer = NULL;
+				juce::AudioBuffer<float> *resampled_buffer = NULL;
 
 				// SLOW DOWN audio (split audio)
-				samples = new juce::AudioSampleBuffer(channels, number_of_samples);
+				samples = new juce::AudioBuffer<float>(channels, number_of_samples);
 				samples->clear();
 
 				// Loop through channels, and get audio samples
@@ -561,7 +563,7 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 																		 reader->info.channels);
 
 					// Allocate a new sample buffer for these delta frames
-					samples = new juce::AudioSampleBuffer(channels, total_delta_samples);
+					samples = new juce::AudioBuffer<float>(channels, total_delta_samples);
 					samples->clear();
 
 					// Loop through each frame in this delta
@@ -569,7 +571,7 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 						 delta_frame <= new_frame_number; delta_frame++) {
 						// buffer to hold detal samples
 						int number_of_delta_samples = GetOrCreateFrame(delta_frame)->GetAudioSamplesCount();
-						juce::AudioSampleBuffer *delta_samples = new juce::AudioSampleBuffer(channels,
+						juce::AudioBuffer<float> *delta_samples = new juce::AudioSampleBuffer(channels,
 																					   number_of_delta_samples);
 						delta_samples->clear();
 
@@ -605,7 +607,7 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 																		 reader->info.channels);
 
 					// Allocate a new sample buffer for these delta frames
-					samples = new juce::AudioSampleBuffer(channels, total_delta_samples);
+					samples = new juce::AudioBuffer<float>(channels, total_delta_samples);
 					samples->clear();
 
 					// Loop through each frame in this delta
@@ -613,7 +615,7 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 						 delta_frame >= new_frame_number; delta_frame--) {
 						// buffer to hold delta samples
 						int number_of_delta_samples = GetOrCreateFrame(delta_frame)->GetAudioSamplesCount();
-						juce::AudioSampleBuffer *delta_samples = new juce::AudioSampleBuffer(channels,
+						juce::AudioBuffer<float> *delta_samples = new juce::AudioSampleBuffer(channels,
 																					   number_of_delta_samples);
 						delta_samples->clear();
 
@@ -644,7 +646,7 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 				resampler->SetBuffer(samples, float(start) / float(number_of_samples));
 
 				// Resample data, and return new buffer pointer
-				juce::AudioSampleBuffer *buffer = resampler->GetResampledBuffer();
+				juce::AudioBuffer<float> *buffer = resampler->GetResampledBuffer();
 
 				// Add the newly resized audio samples to the current frame
 				for (int channel = 0; channel < channels; channel++)
@@ -656,7 +658,7 @@ void Clip::get_time_mapped_frame(std::shared_ptr<Frame> frame, int64_t frame_num
 			}
 			else {
 				// Use the samples on this frame (but maybe reverse them if needed)
-				samples = new juce::AudioSampleBuffer(channels, number_of_samples);
+				samples = new juce::AudioBuffer<float>(channels, number_of_samples);
 				samples->clear();
 
 				// Loop through channels, and get audio samples