--- a/src/CrashHandler.cpp
+++ b/src/CrashHandler.cpp
@@ -12,6 +12,10 @@
 
 #include "CrashHandler.h"
 
+#include <iostream>
+#include <iomanip>
+#include <sstream>
+
 using namespace std;
 using namespace openshot;