… src/DummyReader.cpp
… src/EffectInfo.cpp
… src/Enums.h
… src/FFmpegReader.cpp
… src/FFmpegWriter.cpp
… src/Frame.cpp
… src/Frame.h
… src/FrameMapper.cpp
… src/ImageWriter.cpp
… src/Qt/AudioPlaybackThread.cpp
… src/Qt/AudioPlaybackThread.h
… src/Qt/VideoCacheThread.cpp
… src/Qt/VideoCacheThread.h
… src/Qt/VideoPlaybackThread.cpp
… src/Qt/VideoPlaybackThread.h
… src/QtImageReader.cpp
… src/QtImageReader.h
… src/QtTextReader.cpp
… src/QtTextReader.h
… src/ReaderBase.cpp
… src/ReaderBase.h
… src/Timeline.cpp
… src/WriterBase.cpp
… src/WriterBase.h
… src/ZmqLogger.cpp
… src/ZmqLogger.h
… src/audio_effects/Compressor.cpp
… src/audio_effects/Compressor.h
… src/audio_effects/Delay.cpp
… src/audio_effects/Delay.h