@ferdnyc 2021-11-09
I'd say these numbers speak for themselves.

### Most recent `develop` branch build
![image](https://user-images.githubusercontent.com/538020/140918332-781fabae-08f8-4047-ad4d-0600a82bd287.png)

### This branch
![image](https://user-images.githubusercontent.com/538020/140919252-9f5bdba4-b147-4ae6-8217-76f110cadf08.png)